[ZettaCamp][Module_2] refactor validatepaginationparameters become global validator
