+++ modules/task/task.validator.js
@@ -5,36 +5,7 @@ const { ApolloError } = require('apollo-server');
 // *************** IMPORT VALIDATOR ***************
 const { ValidateMongoId } = require('../../utils/validator/mongo.validator');
 
-/**
- * Validates pagination parameters
- *
- * @function ValidatePaginationParameters
- * @param {Object} params - Parameters object
- * @param {number} params.page - Page number (0-based, where 0 is the first page)
- * @param {number} params.limit - Number of items per page
- * @throws {ApolloError} Throws error if validation fails
- */
-function ValidatePaginationParameters({ page, limit }) {
-  // *************** Check if page is a number
-  if (typeof page !== 'number') {
-    throw new ApolloError('Page must be a number', 'INVALID_PAGINATION');
-  }
-  
-  // *************** Check if page is non-negative
-  if (page < 0) {
-    throw new ApolloError('Page must be a non-negative integer', 'INVALID_PAGINATION');
-  }
-  
-  // *************** Check if limit is a number
-  if (typeof limit !== 'number') {
-    throw new ApolloError('Limit must be a number', 'INVALID_PAGINATION');
-  }
-  
-  // *************** Check if limit is positive
-  if (limit <= 0) {
-    throw new ApolloError('Limit must be a positive integer', 'INVALID_PAGINATION');
-  }
-}
+
 
 /**
  * Validates parameters for creating or updating a task
@@ -146,7 +117,6 @@ function ValidateAssignCorrector(taskId, input) {
 
 // *************** EXPORT MODULE ***************
 module.exports = {
-  ValidatePaginationParameters,
   ValidateCreateUpdateTaskParameters,
   ValidateAssignCorrector
 };
