+++ modules/test/test.resolver.js
@@ -10,6 +10,7 @@ const ErrorLogModel = require('../errorLogs/error_logs.model');
 // *************** IMPORT VALIDATOR ***************
 const TestValidators = require('./test.validator');
 const { ValidateMongoId } = require('../../utils/validator/mongo.validator');
+const { ValidatePaginationParameters } = require('../../utils/validator/pagination.validator');
 
 // *************** QUERY ***************
 /**
@@ -24,8 +25,8 @@ const { ValidateMongoId } = require('../../utils/validator/mongo.validator');
  */
 async function GetAllTests(_, { page, limit }) {
   try {
-    // *************** Validate pagination parameters
-    TestValidators.ValidatePaginationParameters({ page, limit });
+    // *************** Validate pagination parameters 
+    ValidatePaginationParameters({ page, limit });
 
     // *************** Calculate skip value for pagination
     const skip = page * limit;
